IjipMatjip-main/ml-server/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
requests
python-dotenv
joblib
python-multipartlxml
psycopg2-binary
pyproj
